requirements.txt
flask
flask-smorest
python-dotenv
marshmallow
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
